% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_meta.R
\name{ReadMergedExpressTable}
\alias{ReadMergedExpressTable}
\title{Read merged gene expression table}
\usage{
ReadMergedExpressTable(dataName)
}
\arguments{
\item{dataName}{File name of data set.}

\item{norm.opt}{Normalization option (log, vsn, quantile, logCount, NA)}

\item{auto.opt}{integer (0,1), Perform auto scaling}
}
\description{
Perform normalization on individual dataset in meta-analysis
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
